Remove handlers_test.go
