+++ validator/client/beacon-api/BUILD.bazel
@@ -33,7 +33,6 @@ go_test(
         "beacon_api_helpers_test.go",
         "beacon_api_validator_client_test.go",
         "genesis_test.go",
-        "handlers_test.go",
         "json_rest_handler_test.go",
         "wait_for_chain_start_test.go",
     ],
